chore: simplified terraform provider testfixture
[#181898458]
